Fix gson serde compatibility issues with newer jvms (#1)
Use custom type adapter for ByteBuffer json serde

In order to convert `AttributeValue` in newer JVMs that restrict
access to protected modules create a gson type adapter for
`ByteBuffer`s.

The adapter should be able to import "legacy" encoded ByteBuffers
(i.e. those encoded with internal ByteBuffer fields and reflection)
and then use the cleaner formatting of encoding this as base64 strings
in the json (instead of an array of doubles, and other internal fields).